测试了一遍用户端的api

+++ src/main/java/com/example/webmusic/controller/song/SongController.java
@@ -81,7 +81,7 @@ public class SongController {
 
     //根据songid获得一首歌曲
     @GetMapping("/detail/song")
-    public OutApiGetOneSong getOneSong(@RequestParam("songId") int songID) {
+    public OutApiGetOneSong getOneSong(@RequestParam("songId") long songID) {
         OutApiGetOneSong out = new OutApiGetOneSong();
         songService.getOneSongById(songID,out);
         return out;
